% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkainu-package.R
\docType{package}
\name{rkainu-package}
\alias{rkainu}
\alias{rkainu-package}
\title{rkainu: Finnish Spirometry and DLCO Reference Values}
\description{
Predicted values, z-scores, lower limits of normal (LLN), and percent
predicted for Finnish adult spirometry (Kainu et al. 2015) and diffusing
capacity (DLCO; Kainu et al. 2017).
}
\section{Spirometry functions}{

\describe{
\item{\code{\link[=kainu_pred]{kainu_pred()}}}{Predicted (mean) values}
\item{\code{\link[=kainu_lln]{kainu_lln()}}}{Lower limit of normal (5th percentile)}
\item{\code{\link[=kainu_zscore]{kainu_zscore()}}}{Z-scores from observed values}
\item{\code{\link[=kainu_pctpred]{kainu_pctpred()}}}{Percent predicted from observed values}
\item{\code{\link[=kainu_df]{kainu_df()}}}{Data.frame convenience wrapper}
}
}

\section{DLCO functions}{

\describe{
\item{\code{\link[=dlco_pred]{dlco_pred()}}}{Predicted DLCO, DLCO/VA, or VA}
\item{\code{\link[=dlco_lln]{dlco_lln()}}}{Lower limit of normal for DLCO parameters}
\item{\code{\link[=dlco_zscore]{dlco_zscore()}}}{Z-scores from observed DLCO values}
}
}

\references{
Kainu A, Timonen KL, Toikka J, et al. Reference values of spirometry for
Finnish adults. \emph{Clin Physiol Funct Imaging}. 2016;36(5):346-358.
\doi{10.1111/cpf.12237}

Kainu A, Timonen KL, Toikka J, et al. Reference values of diffusing
capacity for Finnish adults. \emph{Clin Physiol Funct Imaging}.
2018;38(3):413-420. \doi{10.1111/cpf.12432}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rikhardfi/rkainu}
  \item Report bugs at \url{https://github.com/rikhardfi/rkainu/issues}
}

}
\author{
\strong{Maintainer}: Rikhard Mäki-Heikkilä \email{rikhard.maki-heikkila@tuni.fi}

}
\keyword{internal}
